backend/requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Database & ORM  
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7

# Background Tasks - FIXED VERSIONS
redis>=4.5.2,<5.3.0
celery>=5.5.2,<5.6.0
kombu>=5.5.2,<5.6.0

# Data Science & ML
numpy>=1.21.0,<1.25.0
pandas>=1.5.0,<2.1.0
scipy>=1.9.0,<1.12.0

# Technical Analysis
ta==0.10.2

# Trading APIs
ccxt==4.1.28

# Machine Learning
scikit-learn>=1.3.0,<1.4.0
joblib==1.3.2

# HTTP & WebSocket
requests==2.31.0
websocket-client==1.6.4
websockets==12.0
aiohttp==3.9.1

# Configuration & Logging
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7

# Enhanced Logging
loguru==0.7.2
coloredlogs==15.0.1

# Security & Crypto
cryptography>=41.0.0,<47.0.0
PyJWT==2.8.0

# Market Data
yfinance==0.2.28

# Development & Testing  
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-mock>=3.12.0,<4.0.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
psutil==5.9.6
orjson==3.9.10
colorama==0.4.6
